Rename main addon
