+++ package.json
@@ -10,12 +10,12 @@
     "test-app"
   ],
   "scripts": {
-    "build": "yarn workspace @ember-responsive-image/core run build && yarn workspace @ember-responsive-image/webpack run build && yarn workspace @ember-responsive-image/blurhash run build && yarn workspace @ember-responsive-image/cloudinary run build && yarn workspace @ember-responsive-image/imgix run build",
+    "build": "yarn workspace ember-responsive-image run build && yarn workspace @ember-responsive-image/webpack run build && yarn workspace @ember-responsive-image/blurhash run build && yarn workspace @ember-responsive-image/cloudinary run build && yarn workspace @ember-responsive-image/imgix run build",
     "lint": "yarn workspaces run lint",
     "lint:fix": "yarn workspaces run lint:fix",
     "prepare": "yarn build",
     "start": "concurrently 'npm:start:*' --restart-after 5000 --prefix-colors cyan,white,yellow",
-    "start:addon": "yarn workspace @ember-responsive-image/core run start",
+    "start:addon": "yarn workspace ember-responsive-image run start",
     "start:test": "yarn workspace test-app run start",
     "test": "yarn workspaces run test",
     "postinstall": "patch-package"
