+++ packages/ember-responsive-image/.eslintrc.js
