+++ packages/ember-responsive-image/.gitignore
