+++ packages/ember-responsive-image/.template-lintrc.js
