+++ packages/ember-responsive-image/addon-main.js
