+++ packages/ember-responsive-image/babel.config.json
