+++ packages/ember-responsive-image/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ember-responsive-image/core",
+  "name": "ember-responsive-image",
   "version": "3.4.0",
   "description": "Automatically generate resized images at build-time, optimized for the responsive web, and using components to render them easily as <picture> elements.",
   "keywords": [
