+++ packages/ember-responsive-image/rollup.config.mjs
