+++ packages/ember-responsive-image/src/components/responsive-image.hbs
