+++ packages/ember-responsive-image/src/helpers/responsive-image-resolve.ts
