+++ packages/ember-responsive-image/src/index.ts
