+++ packages/ember-responsive-image/src/services/responsive-image.ts
