+++ packages/ember-responsive-image/src/template-registry.ts
