+++ packages/ember-responsive-image/src/types.ts
