+++ packages/ember-responsive-image/src/utils/utils.ts
