+++ packages/ember-responsive-image/types/glint.d.ts
