+++ packages/imgix/package.json
@@ -32,7 +32,7 @@
   "peerDependencies": {
     "@glimmer/component": "^1.1.2",
     "@glimmer/tracking": "^1.1.2",
-    "@ember-responsive-image/core": "^3.4.0"
+    "ember-responsive-image": "^3.4.0"
   },
   "devDependencies": {
     "@babel/core": "7.17.0",
@@ -42,7 +42,7 @@
     "@babel/plugin-transform-runtime": "^7.19.6",
     "@babel/preset-env": "7.20.2",
     "@babel/preset-typescript": "7.18.6",
-    "@ember-responsive-image/core": "^3.4.0",
+    "ember-responsive-image": "^3.4.0",
     "@embroider/addon-dev": "2.0.0",
     "@glimmer/component": "1.1.2",
     "@glimmer/tracking": "1.1.2",
