+++ packages/webpack/package.json
@@ -31,7 +31,7 @@
   "devDependencies": {
     "@babel/core": "7.17.0",
     "@babel/preset-typescript": "7.18.6",
-    "@ember-responsive-image/core": "3.4.0",
+    "ember-responsive-image": "3.4.0",
     "@tsconfig/ember": "1.0.0",
     "@types/sharp": "0.31.0",
     "@typescript-eslint/eslint-plugin": "5.30.5",
