+++ packages/webpack/src/images.ts
@@ -1,4 +1,4 @@
-import { ImageType } from '@ember-responsive-image/core/types';
+import { ImageType } from 'ember-responsive-image/types';
 import type { Metadata, Sharp } from 'sharp';
 import type { LoaderContext } from 'webpack';
 import {
