+++ packages/webpack/src/types.ts
@@ -3,7 +3,7 @@ import type {
   LqipBlurhash,
   LqipColor,
   LqipInline,
-} from '@ember-responsive-image/core/types';
+} from 'ember-responsive-image/types';
 import { Metadata, Sharp } from 'sharp';
 
 export type OutputImageType = 'original' | ImageType;
