+++ test-app/package.json
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@ember-responsive-image/blurhash": "3.4.0",
     "@ember-responsive-image/cloudinary": "3.4.0",
-    "@ember-responsive-image/core": "3.4.0",
+    "ember-responsive-image": "3.4.0",
     "@ember-responsive-image/imgix": "3.4.0",
     "@ember-responsive-image/webpack": "3.4.0",
     "@ember/optional-features": "2.0.0",
