+++ test-app/tests/integration/components/responsive-image-test.ts
@@ -3,8 +3,8 @@ import { render, settled } from '@ember/test-helpers';
 import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import { module, test } from 'qunit';
-import { ProviderResult } from '@ember-responsive-image/core/types';
-import ResponsiveImageService from '@ember-responsive-image/core/services/responsive-image';
+import { ProviderResult } from 'ember-responsive-image/types';
+import ResponsiveImageService from 'ember-responsive-image/services/responsive-image';
 import testImage from 'test-app/images/tests/image.jpg?widths=50,100,640&formats=original,webp,avif&responsive';
 import testImageLqipInline from 'test-app/images/tests/image.jpg?lqip=inline&widths=50,100,640&responsive';
 import testImageLqipColor from 'test-app/images/tests/image.jpg?lqip=color&widths=50,100,640&responsive';
